Event Certificates
I am excited to share that my certificates from the recent Event Symposium are now available on GitHub! You can find them in the certificates directory of my repository. If you're interested in viewing them, feel free to check them out here. I had an incredible experience participating in the symposium and am grateful for the opportunity to learn and grow. Thank you to all the organizers and speakers who made this event possible!